SRS document
These are the initial 8 requirements I came up with.  Feel free to add more as needed.  Later on, we will work on implementing these requirements on the development branch by creating "Features" to implement these requirements.  Once these features are working, they can be merged to the development branch in preparation for integration testing.